% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{burritos_cost}
\alias{burritos_cost}
\title{Burrito cost calculator}
\usage{
burritos_cost(price, count)
}
\arguments{
\item{price}{Price per burrito in dollars}

\item{count}{Number of burritos in the order}
}
\description{
Burrito cost calculator
}
