% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal_set_cost.R
\name{meal_set_cost}
\alias{meal_set_cost}
\title{Meal Set Cost Calculator}
\usage{
meal_set_cost(set_name, count)
}
\arguments{
\item{set_name}{"A" or "B"}

\item{count}{Number of meal sets of your specified type that you'd like to purchase}
}
\description{
Meal Set Cost Calculator
}
\examples{
Meal Set A: Tonkatsu (with sauce) over rice with unlimited refills for sides of coleslaw and small soup of the day. Meal Set B: Your choice of grilled chicken or baked tofu, tossed with soy-garlic glaze, sweet corn, and zucchini over rice. Optional: crispy dried seaweed flakes and sesame seeds. Also comes with melted mozzarella cheese on top.
}
