% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy.R
\name{msy}
\alias{msy}
\title{Maximum Sustainable Yield}
\usage{
msy(k, r)
}
\arguments{
\item{k}{Carrying capacity}

\item{r}{Intrinsic growth rate}
}
\description{
Maximum Sustainable Yield
}
\examples{
Calculating sustainable harvest for a fishery
}
