% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushiBear-package.R
\docType{package}
\name{sushiBear-package}
\alias{sushiBear}
\alias{sushiBear-package}
\title{sushiBear: EDS221 Package-building practice}
\description{
Includes several example functions from EDS 221 Day 5 in-class activities, as well as a "Meal Set Calculator" designed by Anna Ramji.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a-ramji/sushiBear}
  \item Report bugs at \url{https://github.com/a-ramji/sushiBear/issues}
}

}
\keyword{internal}
